+++ 05-JWT-Basics/starter/errors/bad-request.js
@@ -1,0 +1,10 @@
+const CustomAPIError = require("./custom-error");
+
+class BadRequest extends CustomAPIError {
+  constructor(message) {
+    super(message);
+    this.statusCode = 400;
+  }
+}
+
+module.exports = BadRequest;
