+++ 05-JWT-Basics/starter/errors/unauthenticated.js
@@ -1,0 +1,10 @@
+const CustomAPIError = require("./custom-error");
+
+class Unauthenticated extends CustomAPIError {
+  constructor(message) {
+    super(message);
+    this.statusCode = 400;
+  }
+}
+
+module.exports = Unauthenticated;
